requirements.txt
backtrader
fugle-trade
fugle-marketdata
python-dotenv
fastapi
pydantic